% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchaeoChron-package.R
\docType{package}
\name{ArchaeoChron-package}
\alias{ArchaeoChron-package}
\alias{ArchaeoChron}
\title{ArchaeoChron: Bayesian Modeling of Archaeological Chronologies}
\description{
Provides a list of functions for the Bayesian modeling of archaeological chronologies. The Bayesian models are implemented in 'JAGS' (Plummer 2003). The inputs are measurements with their associated standard deviations and the study period. The output is the MCMC sample of the posterior distribution of the event date with or without radiocarbon calibration.
}
\details{
\tabular{ll}{
  **Version** \tab 0.2 \cr
  **License** \tab GPL-3 \cr
  **CRAN DOI** \tab \doi{10.32614/CRAN.package.ArchaeoChron} \cr
 }

 Laboratoire de Mathématiques Jean Leray (UMR 6629)\cr
 2, rue de la Houssinière\cr
 BP 92208\cr
 F-44322 Nantes Cedex 3\cr
 France
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ArchaeoStat.github.io/ArchaeoChron/}
  \item \url{https://github.com/ArchaeoStat/ArchaeoChron}
  \item Report bugs at \url{https://github.com/ArchaeoStat/ArchaeoChron/issues}
}

}
\author{
\strong{Maintainer}: Anne Philippe \email{anne.philippe@univ-nantes.fr} (\href{https://orcid.org/0000-0002-5331-5087}{ORCID})

Authors:
\itemize{
  \item Marie-Anne Vibet (\href{https://orcid.org/0000-0003-4003-3141}{ORCID})
}

Other contributors:
\itemize{
  \item Nicolas Frerebeau \email{nicolas.frerebeau@u-bordeaux-montaigne.fr} (\href{https://orcid.org/0000-0001-5759-4944}{ORCID}) [contributor]
  \item Nantes Université (\href{https://ror.org/03gnr7b55}{ROR}) [funder]
  \item CNRS (\href{https://ror.org/02feahw73}{ROR}) [funder]
}

}
\keyword{internal}
