% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intcal13}
\alias{intcal13}
\title{IntCal13}
\format{
An object of class \code{data.frame} with 5141 rows and 3 columns.
}
\usage{
intcal13
}
\description{
IntCal13
}
\seealso{
Other datasets: 
\code{\link{cuers}},
\code{\link{marine13}},
\code{\link{shcal13}},
\code{\link{sunspot}}
}
\concept{datasets}
\keyword{datasets}
