% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marine13}
\alias{marine13}
\title{Marine13}
\format{
An object of class \code{data.frame} with 4801 rows and 3 columns.
}
\usage{
marine13
}
\description{
Marine13
}
\seealso{
Other datasets: 
\code{\link{cuers}},
\code{\link{intcal13}},
\code{\link{shcal13}},
\code{\link{sunspot}}
}
\concept{datasets}
\keyword{datasets}
