% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shcal13}
\alias{shcal13}
\title{ShCal13}
\format{
An object of class \code{data.frame} with 5141 rows and 3 columns.
}
\usage{
shcal13
}
\description{
ShCal13
}
\seealso{
Other datasets: 
\code{\link{cuers}},
\code{\link{intcal13}},
\code{\link{marine13}},
\code{\link{sunspot}}
}
\concept{datasets}
\keyword{datasets}
