% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sunspot}
\alias{sunspot}
\title{Sunspot}
\format{
A [`data.frame`] with 171 rows and 2 columns.
}
\usage{
sunspot
}
\description{
Sunspot
}
\seealso{
Other datasets: 
\code{\link{cuers}}
}
\concept{datasets}
\keyword{datasets}
