% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchaeoChron-package.R
\docType{package}
\name{ArchaeoChron-package}
\alias{ArchaeoChron-package}
\alias{_PACKAGE}
\alias{ArchaeoChron}
\title{ArchaeoChron: Bayesian Modeling of Archaeological Chronologies}
\description{
Provides a list of functions for the Bayesian modeling of archaeological chronologies. The Bayesian models are implemented in 'JAGS' (Plummer 2003). The inputs are measurements with their associated standard deviations and the study period. The output is the MCMC sample of the posterior distribution of the event date with or without radiocarbon calibration.
}
\details{
\tabular{ll}{
  **Package:** \tab ArchaeoChron \cr
  **Type:** \tab Package \cr
  **Version:** \tab 0.2 \cr
  **License:** \tab GPL-3 \cr
 }
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ArchaeoStat.github.io/ArchaeoChron/}
  \item \url{https://github.com/ArchaeoStat/ArchaeoChron}
  \item Report bugs at \url{https://github.com/ArchaeoStat/ArchaeoChron/issues}
}

}
\author{
**Full list of authors and contributors** (alphabetic order)

 \tabular{ll}{
  Nicolas Frerebeau \tab *Université Bordeaux Montaigne, France* \cr
  Anne Philippe \tab *Université de Nantes, France* \cr
  Marie-Anne Vibet \tab *Université de Nantes, France* \cr
 }

 **Package maintainer**

 Anne Philippe\cr
 \email{anne.philippe@univ-nantes.fr}

 Laboratoire de Mathématiques Jean Leray (UMR 6629)\cr
 2, rue de la Houssinière\cr
 BP 92208\cr
 F-44322 Nantes Cedex 3\cr
 France
}
\keyword{internal}
