% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cuers}
\alias{cuers}
\title{Cuers}
\format{
A [`data.frame`] with 2 rows and 3 columns.
}
\usage{
cuers
}
\description{
Cuers
}
\seealso{
Other datasets: 
\code{\link{sunspot}}
}
\concept{datasets}
\keyword{datasets}
