% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventModel_C14.R
\name{eventModel_C14}
\alias{eventModel_C14}
\title{Event Model for Radiocarbon Dates}
\usage{
eventModel_C14(
  M,
  s,
  studyPeriodMin,
  studyPeriodMax,
  calibCurve = "intcal13",
  numberChains = 2,
  numberAdapt = 10000,
  numberUpdate = 10000,
  variable.names = c("theta"),
  numberSample = 50000,
  thin = 10
)
}
\arguments{
\item{M}{A [`numeric`] vector of measurements.}

\item{s}{A [`numeric`] vector of errors.}

\item{studyPeriodMin}{A length-one [`numeric`] vector specifying the start
time of the study period.}

\item{studyPeriodMax}{A length-one [`numeric`] vector specifying the end
time of the study period.}

\item{calibCurve}{A [`character`] string specifying the calibration curve to
be used.}

\item{numberChains}{An [`integer`] giving the number of of parallel chains
for the model (see [rjags::jags.model()]).}

\item{numberAdapt}{An [`integer`] giving the number of iterations for
adaptation (see [rjags::jags.model()]).}

\item{numberUpdate}{An [`integer`] giving the number of iterations to update
the model by.}

\item{variable.names}{A [`character`] vector giving the names of variables to
be monitored (see [rjags::coda.samples()]).}

\item{numberSample}{An [`integer`] giving the number of iterations to
monitor (see [rjags::coda.samples()]).}

\item{thin}{An [`integer`] giving the thinning interval for monitors
(see [rjags::coda.samples()]).}
}
\value{
An [`mcmc.list`][coda::mcmc.list()] object.
}
\description{
Event Model for Radiocarbon Dates
}
\examples{
\donttest{
eve <- eventModel_C14(
  M = 4500,
  s = 35,
  studyPeriodMin = 700,
  studyPeriodMax = 1100,
  calibCurve = "intcal20",
  numberSample = 5000
)
plot(eve)
}
}
\author{
A. Philippe, M.-A. Vibet
}
