\name{shcal13}
\alias{shcal13}
\docType{data}
\title{
Southern Hemisphere Atmospheric Radiocarbon Calibration Curve
}
\description{
Southern Hemisphere atmospheric radiocarbon calibration curve published by Reimer et al. 
}
\usage{data("shcal13")}
\format{
  A data frame with 5141 observations on the following 3 variables.
  \describe{
    \item{\code{CALBP}}{a numeric vector of calibrated age in year before present (before 1950)}
    \item{\code{14Cage}}{a numeric vector of radiocarbon age in year before present (before 1950)}
    \item{\code{Error}}{a numeric vector of calibrated age in year before present (before 1950)}
  }
}
\references{
Reimer PJ, Bard E, Bayliss A, Beck JW, Blackwell PG, Bronk Ramsey C, Buck CE, Cheng H, Edwards RL, Friedrich M, Grootes PM, Guilderson TP, Haflidason H, Hajdas I, Hatte C, Heaton TJ, Hoffmann DL, Hogg AG, Hughen KA, Kaiser KF, Kromer B, Manning SW, Niu M, Reimer RW, Richards DA, Scott EM, Southon JR, Staff RA, Turney CSM, van der Plicht J. 2013. IntCal13 and Marine13 radiocarbon age calibration curves 0-50,000 years cal BP. Radiocarbon 55(4):1869-1887.
}
\examples{
data(shcal13)
}
\keyword{datasets}
