% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sunspot}
\alias{sunspot}
\title{Sunspot}
\format{
An object of class \code{data.frame} with 171 rows and 2 columns.
}
\usage{
sunspot
}
\description{
Sunspot
}
\seealso{
Other datasets: 
\code{\link{cuers}},
\code{\link{intcal13}},
\code{\link{marine13}},
\code{\link{shcal13}}
}
\concept{datasets}
\keyword{datasets}
