\name{sunspot}
\alias{sunspot}
\docType{data}
\title{
Dated impacts on the sun corresponding to a unique archaeological event (Fictive data)
}
\description{
Fictive data corresponding to dated impacts on the sun
}
\usage{data("sunspot")}
\format{
  A data frame with 171 observations on the following 2 variables.
  \describe{
    \item{\code{Age}}{a numeric vector corresponding to the dated impact (ages in date before 2016)}
    \item{\code{Error}}{a numeric vector corresponding to the error made on the measurement}
  }
}
\examples{
data(sunspot)
}
\keyword{datasets}
